+++ app/src/main/java/com/proyectofct/salinappservice/Clases/Reservas/ListaProductosCarritoAdapter.java
@@ -1,25 +1,41 @@
 package com.proyectofct.salinappservice.Clases.Reservas;
 
 import android.content.Context;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.OnFailureListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.firestore.DocumentReference;
+import com.google.firebase.firestore.DocumentSnapshot;
+import com.google.firebase.firestore.FieldValue;
+import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.firestore.SetOptions;
 import com.proyectofct.salinappservice.Clases.Productos.ProductoCarrito;
 import com.proyectofct.salinappservice.R;
 
 import org.jetbrains.annotations.NotNull;
 
 import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Map;
 
 public class ListaProductosCarritoAdapter extends RecyclerView.Adapter<ProductoCarritoViewHolder> {
     private Context c;
     private ArrayList<ProductoCarrito> listaProductosCarrito;
     private LayoutInflater inflater;
+
+    private FirebaseAuth firebaseAuth;
+    private FirebaseFirestore db;
 
     public ListaProductosCarritoAdapter(Context c, ArrayList<ProductoCarrito> listaProductosCarrito) {
         this.c = c;
@@ -47,6 +63,8 @@ public class ListaProductosCarritoAdapter extends RecyclerView.Adapter<ProductoC
     @Override
     public ProductoCarritoViewHolder onCreateViewHolder(@NonNull @NotNull ViewGroup parent, int viewType) {
         View item = inflater.inflate(R.layout.item_recyclerview_producto_carrito, parent, false);
+        db = FirebaseFirestore.getInstance();
+        firebaseAuth = FirebaseAuth.getInstance();
         return new ProductoCarritoViewHolder(item, this);
     }
 
@@ -62,6 +80,31 @@ public class ListaProductosCarritoAdapter extends RecyclerView.Adapter<ProductoC
         holder.txtMarcaProductoCarrito.setText("Marca: " + String.valueOf(productoCarritoActual.getMarca()));
         holder.txtCantidadProductoCarrito.setText("Cantidad: " + String.valueOf(productoCarritoActual.getCantidad()) + " unidades");
         holder.txtPrecioProductoCarrito.setText("Precio :" + String.valueOf(productoCarritoActual.getPrecio()) + "€");
+
+        //AUMENTAR O DISMINUIR CANTIDAD DEL PRODUCTO
+        holder.btAumentarCantidad.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                productoCarritoActual.setCantidad(productoCarritoActual.getCantidad() + 1);
+                aumentarCantidadProductoFirestore(productoCarritoActual);
+                cargarRecyclerView(productoCarritoActual);
+            }
+        });
+
+        holder.btDisminuirCantidad.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                if (productoCarritoActual.getCantidad() <= 1){
+                    productoCarritoActual.setCantidad(productoCarritoActual.getCantidad() - 1);
+                    borrarProductoFirestore(productoCarritoActual);
+                    cargarRecyclerView(productoCarritoActual);
+                }else if(productoCarritoActual.getCantidad() > 1){
+                    productoCarritoActual.setCantidad(productoCarritoActual.getCantidad() - 1);
+                    disminuirCantidadProductoFirestore(productoCarritoActual);
+                    cargarRecyclerView(productoCarritoActual);
+                }
+            }
+        });
     }
 
     @Override
@@ -77,5 +120,87 @@ public class ListaProductosCarritoAdapter extends RecyclerView.Adapter<ProductoC
     public void cargarRecyclerView(ProductoCarrito productoCarrito) {
         listaProductosCarrito.add(productoCarrito);
         notifyItemInserted(listaProductosCarrito.size());
+    }
+
+    //MÉTODO PARA BORRAR EL PRODUCTO DE FIRESTORE
+    public void borrarProductoFirestore(ProductoCarrito productoCarrito) {
+        DocumentReference documentReference = db.collection("shoppingcars").document(firebaseAuth.getCurrentUser().getUid());
+        Map<String, Object> productosCarrito = new HashMap<>();
+        productosCarrito.put(String.valueOf(productoCarrito.getCodProducto()), FieldValue.delete());
+
+        documentReference.update(productosCarrito).addOnCompleteListener(new OnCompleteListener<Void>() {
+            @Override
+            public void onComplete(@NonNull @org.jetbrains.annotations.NotNull Task<Void> task) {
+                Toast.makeText(getC(), "Producto eliminado del carrito correctamente", Toast.LENGTH_SHORT).show();
+            }
+        });
+
+        documentReference.update(productosCarrito).addOnFailureListener(new OnFailureListener() {
+            @Override
+            public void onFailure(@NonNull @org.jetbrains.annotations.NotNull Exception e) {
+                Log.i("", "Error al borrar el producto del carrito");
+            }
+        });
+    }
+
+    //MÉTODO PARA AUMENTAR LA CANTIDAD DEL PRODUCTO DE FIRESTORE
+    public void aumentarCantidadProductoFirestore(ProductoCarrito productoCarrito){
+        DocumentReference documentReference = db.collection("shoppingcars").document(firebaseAuth.getCurrentUser().getUid());
+        documentReference.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
+            @Override
+            public void onComplete(@NonNull @com.google.firebase.database.annotations.NotNull Task<DocumentSnapshot> task) {
+                if(task.isComplete()){
+                    task.getResult();
+                    DocumentSnapshot data = task.getResult();
+                    ProductoCarrito miProductoCarrito = data.get(String.valueOf(productoCarrito.getCodProducto()), ProductoCarrito.class);
+                    if(miProductoCarrito != null){
+                        int nuevaCantidad = miProductoCarrito.getCantidad() + 1;
+                        miProductoCarrito.setCantidad(nuevaCantidad);
+                        actualizarProductoFirestore(miProductoCarrito, documentReference);
+                    }else{
+                        actualizarProductoFirestore(productoCarrito, documentReference);
+                    }
+                }
+            }
+        });
+    }
+
+    //MÉTODO PARA DISMINUIR LA CANTIDAD DEL PRODUCTO DE FIRESTORE
+    public void disminuirCantidadProductoFirestore(ProductoCarrito productoCarrito){
+        DocumentReference documentReference = db.collection("shoppingcars").document(firebaseAuth.getCurrentUser().getUid());
+        documentReference.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
+            @Override
+            public void onComplete(@NonNull @com.google.firebase.database.annotations.NotNull Task<DocumentSnapshot> task) {
+                if(task.isComplete()){
+                    task.getResult();
+                    DocumentSnapshot data = task.getResult();
+                    ProductoCarrito miProductoCarrito = data.get(String.valueOf(productoCarrito.getCodProducto()), ProductoCarrito.class);
+                    if(miProductoCarrito != null){
+                        int nuevaCantidad = miProductoCarrito.getCantidad() - 1;
+                        miProductoCarrito.setCantidad(nuevaCantidad);
+                        actualizarProductoFirestore(miProductoCarrito, documentReference);
+                    }else{
+                        actualizarProductoFirestore(productoCarrito, documentReference);
+                    }
+                }
+            }
+        });
+    }
+
+    public void actualizarProductoFirestore(ProductoCarrito productoCarrito, DocumentReference documentReference){
+        Map<String, Object> mapaProductoCarrito = new HashMap<>();
+        mapaProductoCarrito.put(String.valueOf(productoCarrito.getCodProducto()), productoCarrito);
+
+        documentReference.set(mapaProductoCarrito, SetOptions.merge()).addOnCompleteListener(new OnCompleteListener<Void>() {
+            @Override
+            public void onComplete(@NonNull Task<Void> task) {
+                Toast.makeText(getC(), "Producto actualizado correctamente en Firestore", Toast.LENGTH_LONG).show();
+            }
+        }).addOnFailureListener(new OnFailureListener() {
+            @Override
+            public void onFailure(@NonNull Exception e) {
+                Log.i("","Error al actualizar el producto en Firestore");
+            }
+        });
     }
 }
