+++ app/src/main/res/layout/item_recyclerview_producto_carrito.xml
@@ -57,11 +57,11 @@
             <Button
                 android:id="@+id/btAumentarCantidad"
                 android:layout_gravity="left"
-                android:layout_width="95dp"
+                android:layout_width="90dp"
                 android:layout_height="50dp"
                 android:layout_marginTop="10dp"
                 android:text="@string/aumentar_cantidad"
-                android:textSize="10sp"
+                android:textSize="9.5sp"
                 android:textColor="@color/black"
                 android:backgroundTint="@color/fondo"
                 app:cornerRadius="25dip" />
@@ -70,10 +70,10 @@
                 android:id="@+id/btDisminuirCantidad"
                 android:layout_gravity="right"
                 android:layout_marginTop="-50dp"
-                android:layout_width="95dp"
+                android:layout_width="90dp"
                 android:layout_height="50dp"
                 android:text="@string/disminuir_cantidad"
-                android:textSize="10sp"
+                android:textSize="9.5sp"
                 android:textColor="@color/black"
                 android:backgroundTint="@color/fondo"
                 app:cornerRadius="25dip" />
